fix listing of tracks for managed users
